requirements.txt
easygui==0.98.1
requests==2.22.0
praw==7.1.0
art==4.8
Pillow==7.2.0
